+++ web/src/components/foundation/index.ts
@@ -3,3 +3,4 @@ export * from './CodeText';
 export * from './StatusText';
 export * from './Divider';
 export * from './VisuallyHidden';
+export * from './Caption';