+++ web/src/components/layout/index.ts
@@ -2,3 +2,7 @@ export * from './Panel';
 export * from './Stack';
 export * from './Inline';
 export * from './DashboardGrid';
+export * from './AppShell';
+export * from './ScrollRegion';
+export * from './TabList';
+export * from './FormRow';