+++ web/src/components/molecules/index.ts
@@ -1,2 +1,4 @@
 export * from './CoveragePanel';
 export * from './QueryPresetList';
+export * from './MetadataGrid';
+export * from './DataTable';